requirements.txt
google-cloud-aiplatform==1.69.0